% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateModels.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluate orthogonal neural additive model}
\usage{
evaluateModel(PHOModelList)
}
\arguments{
\item{PHOModelList}{Orthogonal neural additive model ensemble object to be evaluated}
}
\description{
Evaluate orthogonal neural additive model
}
