% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotMainEffect}
\alias{plotMainEffect}
\title{Plot Main Effect}
\usage{
plotMainEffect(evalData, effect)
}
\arguments{
\item{evalData}{Model output as obtained from ONAM::evaluateModel}

\item{effect}{Effect to be plotted, must be present in the model formula. For interaction terms, use plotInteractionEffect}
}
\value{
Returns a ggplot2 object of the specified effect
}
\description{
Plot Main Effect
}
